homework-4 is ready

+++ order-service/src/main/java/ru/liga/dto/ItemDto.java
@@ -5,7 +5,10 @@ import lombok.Data;
 @Data
 public class ItemDto {
     private Double price;
-    private Integer quantity;
+
+    private Long quantity;
+
     private String description;
+
     private String image;
 }
